requirements.txt
requests
fastapi
uvicorn
python-multipart
GitPython
